refactor existing tests
